AC-15620: Ensure PHP 8.5 Compatibility with adobe commerce

+++ app/code/Magento/CatalogInventory/Model/ResourceModel/Indexer/Stock/DefaultStock.php
@@ -333,7 +333,7 @@ class DefaultStock extends AbstractIndexer implements StockInterface
                 'product_id' => (int)$row['entity_id'],
                 'website_id' => (int)$row['website_id'],
                 'stock_id' => Stock::DEFAULT_STOCK_ID,
-                'qty' => (double)$row['qty'],
+                'qty' => (float)$row['qty'],
                 'stock_status' => (int)$row['status'],
             ];
             $savedEntityIds[] = (int)$row['entity_id'];
